+++ src/main/java/com/heeverse/config/MultiDataSource.java
@@ -5,21 +5,25 @@ import com.zaxxer.hikari.HikariDataSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
+import org.springframework.core.env.Environment;
+import org.springframework.util.Assert;
 
-import static com.heeverse.common.Constants.MYSQL_SECRETES;
-import static com.heeverse.common.Constants.VAULT_PATH;
+import java.util.Arrays;
+
+import static com.heeverse.common.Constants.*;
 
 /**
  * @author jeongheekim
  * @date 10/31/23
  */
 @Configuration
-@MultiDataSourceProfile
 public class MultiDataSource {
     private final VaultOperationService vaultOperationService;
+    private final Environment environment;
 
-    public MultiDataSource(VaultOperationService vaultOperationService) {
+    public MultiDataSource(VaultOperationService vaultOperationService, Environment environment) {
         this.vaultOperationService = vaultOperationService;
+        this.environment = environment;
     }
 
     @Primary
@@ -34,12 +38,34 @@ public class MultiDataSource {
     }
 
     private HikariConfig getDataSourceProperties() {
-        var dbProps = vaultOperationService.getProps(VAULT_PATH, MYSQL_SECRETES, RdsConnectionProps.DBProps.class);
+        String[] activeProfiles = environment.getActiveProfiles();
         HikariConfig hikariConfig = new HikariConfig();
-        hikariConfig.setJdbcUrl(dbProps.url());
-        hikariConfig.setUsername(dbProps.username());
-        hikariConfig.setPassword(dbProps.password());
+        if (Arrays.asList(activeProfiles).contains(LOCAL)) {
+            hikariConfig.setJdbcUrl(environment.getProperty("spring.datasource.url"));
+            hikariConfig.setDriverClassName(environment.getProperty("spring.datasource.driver-class-name"));
+            hikariConfig.setUsername(environment.getProperty("spring.datasource.username"));
+        } else {
+            var dbProps = vaultOperationService.getProps(VAULT_PATH, MYSQL_SECRETES, DBProps.class);
+            hikariConfig.setJdbcUrl(dbProps.url());
+            hikariConfig.setUsername(dbProps.username());
+            hikariConfig.setPassword(dbProps.password());
+        }
+
         return hikariConfig;
+    }
+
+
+    public record DBProps(
+            String url,
+            String username,
+            String password
+    ) {
+        public DBProps {
+            String name = this.getClass().getSimpleName();
+            Assert.notNull(url, name + "'s [url] must not null!");
+            Assert.notNull(username, name + "'s [username] must not null!");
+            Assert.notNull(password, name + "'s [password] must not null!");
+        }
     }
 
 }
