+++ src/main/resources/application-dev-test.yml
@@ -1,16 +1,9 @@
 spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
+  sql:
+    init:
+      mode: never
   jmx:
     enabled: false
-
-mybatis:
-  mapper-locations:
-    - classpath:mapper/basic/*.xml
-    - classpath:mapper/lock-mapper/*.xml
-  configuration:
-    map-underscore-to-camel-case: true
-    arg-name-based-constructor-auto-mapping: true
 
 logging:
   level:
