+++ src/main/resources/application-prod.yml
@@ -2,14 +2,6 @@ spring:
   jmx:
     enabled: false
 
-
-
-mybatis:
-  mapper-locations: classpath:mapper/*.xml
-  configuration:
-    map-underscore-to-camel-case: true
-    arg-name-based-constructor-auto-mapping: true
-
 logging:
   level:
     org:
